Level 1 - Basic UI and Level Setup
Additions:
Added a UI Canvas as well as Text boxes for score Counters.
Added an event system in case we want to have player input on a UI layer before/after transitioning to the next level. This can be disabled during gameplay if it interferes, and reenabled after the level has reached a win/loss state.
Added properly sized Paddles and a Ball with no gravity in the middle of the screen.
Added Upper and lower bounds with 2d Box colliders to keep the ball in.
Added in a Player and Enemy goal that are colored accordingly with box colliders that function are triggers to be used to score points.